requirements.txt
google-adk[database]
python-dotenv